requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
discord.py
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
multidict==6.3.2
propcache==0.3.1
pycparser==2.22
PyNaCl==1.5.0
rfc3986==1.5.0
setuptools==58.1.0
sniffio==1.3.1
typing_extensions==4.13.1
yarl==1.19.0
youtube-search-python==1.6.6
yt-dlp==2025.10.14